requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
curl_cffi==0.6.2
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
frozenlist==1.4.1
g4f==0.2.5.7
gunicorn==21.2.0
h11==0.14.0
idna==3.6
multidict==6.0.5
outcome==1.3.0.post0
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
undetected-chromedriver==3.5.5
urllib3==2.2.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
